--- a/Cargo.toml
+++ b/Cargo.toml
@@ -18,3 +18,7 @@ ctrlc = "3.1"
 [dependencies.clap]
 version = "2"
 features = ["suggestions", "color", "wrap_help"]
+
+[profile.release]
+lto = true
+codegen-units = 1