this PR:

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `hexyl data` | 588.3 ± 3.7 | 585.2 | 596.1 | 36.01 |
| `hexyl --color=never data` | 431.6 ± 7.0 | 425.2 | 447.5 | 26.42 |
| `hexdump -C data` | 16.3 ± 0.6 | 15.7 | 20.2 | 1.00 |
| `xxd data` | 706.3 ± 4.6 | 701.5 | 716.4 | 43.24 |

before:

| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `hexyl data` | 1.031 ± 0.007 | 1.023 | 1.047 | 62.40 |
| `hexyl --color=never data` | 0.839 ± 0.005 | 0.834 | 0.848 | 50.79 |

benchmark command:
```
dd if=/dev/zero    bs=10M count=1 >  data
dd if=/dev/urandom bs=1k  count=1 >> data

hyperfine --warmup 5 \   
    'hexyl data' \
    'hexyl --color=never data' \
    'hexdump -C data' \
    'xxd data' \
    --export-markdown results.md
```

this means that we now clearly beat `xxd` (in this scenario) and can close #66  :smile: